Revert "Revert "Type IngredientDetail Page, components order-info, constructor, detail""
This reverts commit 782531bd6234358a75b615155b05cb3a7d30abc1.
