+++ src/components/burger-constructor/burger-constructor.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState, useMemo } from "react";
+import React, { useEffect, useState, useMemo, FunctionComponent } from "react";
 import { useDrop } from "react-dnd";
 import { useDispatch, useSelector } from "react-redux";
 import { useHistory } from 'react-router-dom';
@@ -7,9 +7,9 @@ import { CurrencyIcon } from '@ya.praktikum/react-developer-burger-ui-components
 import { Button } from '@ya.praktikum/react-developer-burger-ui-components';
 import { ConstructorElement } from '@ya.praktikum/react-developer-burger-ui-components';
 import styles from "./burger-constructor.module.css";
-import LayerElement from "../layer-element/layer-element";
+import { LayerElement } from "../layer-element/layer-element";
 import Modal from '../modal/modal';
-import OrderInfo from "../order-info/order-info";
+import { OrderInfo } from "../order-info/order-info";
 import {
     addIngredient,
     addOrChangeBun,
