+++ src/pages/constructor/constructor.tsx
@@ -1,15 +1,15 @@
 
-import React from 'react';
+import React, { FunctionComponent } from 'react';
 import { useSelector } from 'react-redux';
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
 import styles from './constructor.module.css';
-import BurgerIngredients from '../../components/burger-ingredients/burger-ingredients';
+import { BurgerIngredients } from '../../components/burger-ingredients/burger-ingredients';
 import BurgerConstructor from '../../components/burger-constructor/burger-constructor';
 import Modal from '../../components/modal/modal';
 import { getLoadingStatus } from '../../services/selectors/selectors';
 
-export const ConstructorPage = () => {
+export const ConstructorPage: FunctionComponent = () => {
     const loading = useSelector(getLoadingStatus);
 
     return (
@@ -21,7 +21,7 @@ export const ConstructorPage = () => {
                 </div>
             </DndProvider>
             {loading &&
-                (<Modal title='LOADING...' onClose={() => { }} />)
+                (<Modal title='LOADING...' onClose={() => { } } children={undefined} />)
             }
         </>
     );
